requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiojobs==1.3.0
aiosignal==1.3.1
aiosmtpd==1.4.6
alembic==1.13.2
amqp==5.2.0
apispec==6.6.1
appdirs==1.4.4
apsw==3.46.1.0
asgiref==3.8.1
atpublic==5.0
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.8.2
Bottleneck==1.4.0
branca==0.8.0
Brotli==1.1.0
bs4==0.0.2
cachelib==0.9.0
cachetools==5.5.0
cattrs==24.1.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
crispy-bootstrap4==2024.1
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==42.0.8
cycler==0.12.1
dateparser==1.2.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.0.7
django-admin-tools==0.9.3
django-adminlte-ui==2.1.1
django-appconf==1.0.6
django-baton==4.1.0
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-cryptography-django5==2.2
django-filter==24.3
django-grappelli==4.0.1
django-guardian==2.4.0
django-import-export==4.1.1
django-jazzmin==3.0.0
django-jet==1.0.8
django-js-asset==2.2.0
django-material-admin==1.8.6
django-mathfilters==1.0.0
django-select2==8.2.0
django-semantic-admin==0.6.2
django-semantic-forms==0.1.8
django-simple-history==3.7.0
django-simpleui==2024.11.15
django-slick-reporting==1.3.1
django-sql-explorer==5.3
django-suit==0.2.28
django-tinymce==4.1.0
django-unfold==0.41.0
django-xadmin==0.5.0
dnspython==2.6.1
docx==0.2.4
email_validator==2.2.0
feedparser==6.0.11
Flask==2.3.3
Flask-AppBuilder==4.5.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-Compress==1.15
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
folium==0.18.0
fonttools==4.53.1
frozenlist==1.5.0
func_timeout==4.3.5
geographiclib==2.0
geopy==2.4.1
google-auth==2.34.0
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hashids==1.3.1
holidays==0.25
humanize==4.10.0
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.4.0
korean-lunar-calendar==0.3.1
limits==3.13.0
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.1
mdurl==0.1.2
msgpack==1.0.8
msgspec==0.18.6
multidict==6.1.0
netifaces==0.11.0
nh3==0.2.18
numexpr==2.10.1
numpy==2.0.1
openai==0.28.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
paramiko==3.4.1
parsedatetime==2.6
pgsanity==0.2.9
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
polyline==2.0.2
prison==0.2.1
prompt_toolkit==3.0.47
propcache==0.2.0
psycopg2-binary==2.9.9
pyarrow==18.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-geohash==0.8.5
python-nmap==0.7.1
pytz==2024.1
PyYAML==6.0.2
redis==4.6.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-cache==1.2.1
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
scapy==2.6.1
seaborn==0.13.2
selenium==4.9.1
setuptools==75.5.0
sgmllib3k==1.0.0
shelved-cache==0.3.1
shillelagh==1.2.26
shortid==0.1.2
simplejson==3.19.3
six==1.16.0
slack_sdk==3.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.38.3
sqlglot==23.17.0
sqlparse==0.5.1
sshtunnel==0.4.0
tablib==3.5.0
tabulate==0.8.10
tenacity==9.0.0
tqdm==4.67.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
url-normalize==1.4.3
urllib3==2.2.2
vine==5.1.0
wbdata==1.0.0
wbgapi==1.0.12
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
whitenoise==6.7.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
WTForms-JSON==0.3.5
XlsxWriter==3.0.9
xyzservices==2024.9.0
yarl==1.17.2
zipp==3.20.1
zstandard==0.23.0
html2docx    
django-cors-headers
django_countries
openai
natsort
face_recognition
pdf2image
deepface
tf-keras
fitz
frontend
tools
pyarrow pandas